add two step store deletion e2e case for tiflash
