+++ tests/e2e/e2e_test.go
@@ -29,6 +29,7 @@ import (
 	_ "github.com/pingcap/tidb-operator/tests/e2e/pd"
 	_ "github.com/pingcap/tidb-operator/tests/e2e/ticdc"
 	_ "github.com/pingcap/tidb-operator/tests/e2e/tidb"
+	_ "github.com/pingcap/tidb-operator/tests/e2e/tiflash"
 	_ "github.com/pingcap/tidb-operator/tests/e2e/tikv"
 	_ "github.com/pingcap/tidb-operator/tests/e2e/webhook"
 )
